場中の株価が最大12時間固まる問題を修正
キャッシュの有効期限が投信も株も一律12時間だったため、場中に一度取得
すると、その後半日は同じ株価が表示され続けていた。日付は当日でも値が
動かないので「更新されない」ように見える。

- キャッシュの有効期限を種別で分け、株は取引時間中（平日9:00〜15:40）
  のみ15分にする。投信の基準価額は1日1回なので12時間のまま
- get_cached_series の max_age_hours を float に（15分＝0.25時間のため）
- check_stock.py に「いつ取得したか・有効期限まであと何分か」を表示。
  日付が当日でも値が止まる状態を切り分けられるようにした

+++ fund-timing/db.py
@@ -742,7 +742,7 @@ def trade_position(trades, kind: str = "fund") -> dict:
 
 
 # ------------------------------------------------------------------ cache
-def get_cached_series(isin: str, assoc_code: str, max_age_hours: int = CACHE_TTL_HOURS,
+def get_cached_series(isin: str, assoc_code: str, max_age_hours: float = CACHE_TTL_HOURS,
                       db_path: Optional[str] = None):
     with _conn(db_path) as c:
         r = c.execute(
